+++ src/libgtmatch/test-mergeesa.c
@@ -80,7 +80,7 @@ static int outputsuflcpllv(void *processinfo,
   {
     env_error_set(env,"fwrite(%s) of %u Seqpos-value failed: %s",
                   str_get(mergeoutinfo->outsuf.outfilename),
-                  (unsigned int) buf->nextstoreidx,strerror(errno));
+                  buf->nextstoreidx,strerror(errno));
     haserr = true;
   }
   if (!haserr)
